# cmd/api/docs/swagger.yaml
definitions:
...
  schemas.PermissionResponse:
    properties:
      code:
        type: string
      details:
        type: string
      group:
        type: string
      id:
        type: string
    type: object
...
  schemas.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
  schemas.RoleCreate:
    properties:
      name:
        type: string
      permissions_id:
        items:
          type: string
        type: array
    required:
    - name
    - permissions_id
    type: object
...
  schemas.RoleResponse:
    properties:
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/schemas.PermissionResponse'
        type: array
    type: object
...
info:
  contact: {}
  description: This is a api to app noa gestion
  termsOfService: http://swagger.io/terms/
  title: APP NOA Gestion
  version: "1.0"
paths:
...
  /api/v1/role/create:
    post:
      consumes:
      - application/json
      description: This function fetches roles based on the user's role and workplace
        identifier
      parameters:
      - description: Role object
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/schemas.RoleCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Roles retrieved successfully
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad request if user or workplace is missing
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error on failure
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Retrieve roles for a user in a specific workplace
      tags:
      - Role
  /api/v1/role/get_all:
    get:
      consumes:
      - application/json
      description: This function fetches roles based on the user's role and workplace
        identifier
      produces:
      - application/json
      responses:
        "200":
          description: Roles retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.RoleResponse'
                  type: array
              type: object
        "400":
          description: Bad request if user or workplace is missing
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal server error on failure
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: Retrieve roles for a user in a specific workplace
      tags:
      - Role
...
securityDefinitions:
  CookieAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"